refactor: delete battlefield-grid component
