--- src/components/battlefield-grid/battlefield-grid.scss
@@ -1,20 +1,0 @@
-.battlefield-grid {
-  display: grid;
-
-  grid-template-areas:
-    "topLine topLine topLine topLine topLine topLine topLine topLine topLine topLine topLine"
-    "leftLine cell-00 cell-01 cell-02 cell-03 cell-04 cell-05 cell-06 cell-07 cell-08 cell-09"
-    "leftLine cell-10 cell-11 cell-12 cell-13 cell-14 cell-15 cell-16 cell-17 cell-18 cell-19"
-    "leftLine cell-20 cell-21 cell-22 cell-23 cell-24 cell-25 cell-26 cell-27 cell-28 cell-29"
-    "leftLine cell-30 cell-31 cell-32 cell-33 cell-34 cell-35 cell-36 cell-37 cell-38 cell-39"
-    "leftLine cell-40 cell-41 cell-42 cell-43 cell-44 cell-45 cell-46 cell-47 cell-48 cell-49"
-    "leftLine cell-50 cell-51 cell-52 cell-53 cell-54 cell-55 cell-56 cell-57 cell-58 cell-59"
-    "leftLine cell-60 cell-61 cell-62 cell-63 cell-64 cell-65 cell-66 cell-67 cell-68 cell-69"
-    "leftLine cell-70 cell-71 cell-72 cell-73 cell-74 cell-75 cell-76 cell-77 cell-78 cell-79"
-    "leftLine cell-80 cell-81 cell-82 cell-83 cell-84 cell-85 cell-86 cell-87 cell-88 cell-89"
-    "leftLine cell-90 cell-91 cell-92 cell-93 cell-94 cell-95 cell-96 cell-97 cell-98 cell-99";
-
-  row-gap: 0.1rem;
-  column-gap: 0.1rem;
-  background-color: black;
-}
