--- src/components/battlefield-grid/index.ts
@@ -1,3 +1,0 @@
-import Grid from './battlefield-grid';
-
-export default Grid;
